Fixed timer
Fixed the timer. We don’t need one. time is arbitrary.
we just need to put in a delay each time before updating.
Maybe this.wait(1 second); and the end of each loop.

We also need to make sure that the formulas are correct.
It is still getting stuck in an infinite loop somewhere